Работа над заданием
