+++ index.html
@@ -6,8 +6,10 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/functions.js" type="module"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/util.js" type="module"></script>
+  <script src="js/data.js" type="module"></script>
   <title>Кекстаграм</title>
 </head>
